+++ src/main/java/sogang/capstone/editking/domain/form/FormCatalogService.java
@@ -1,23 +1,8 @@
 package sogang.capstone.editking.domain.form;
 
-import java.util.List;
-import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import sogang.capstone.editking.domain.user.User;
 
-@Service
-@RequiredArgsConstructor
-public class FormCatalogService {
+public interface FormCatalogService {
 
-    private final FormReader formReader;
-
-    @Transactional(readOnly = true)
-    public FormInfo.CatalogMain retrieveFormCatalog(User user, String status, Integer limit) {
-        List<Form> formList = formReader.getFormCatalog(user, FormStatus.valueOf(status), limit);
-        List<FormInfo.CatalogForm> catalogFormList = formList.stream().map(FormInfo.CatalogForm::new)
-                .collect(Collectors.toList());
-        return new FormInfo.CatalogMain(catalogFormList);
-    }
+    public FormInfo.CatalogMain retrieveFormCatalog(User user, String status, Integer limit);
 }
